docs: validate:error-codes target + scripts/lib/ in CLAUDE.md/AGENTS.md (#531)
Address P2 docs-coverage findings from review of #531:

- CLAUDE.md "Common Commands": add `pnpm validate:error-codes` line; update
  `validate:schemas` comment to include the new error-codes pass.
- AGENTS.md "Build, Test, And Development Commands": same updates; describe
  the lint-error-codes.js scope.
- AGENTS.md "Project Structure": document scripts/lib/ — validate-plugin.js
  and validate-marketplace.js are now thin orchestrators that import shared
  helpers from there.
- tests/integration/validate-plugin-characterization.test.ts: update doc
  comment to match actual behavior (errors are sorted before snapshot, so
  rule ordering is NOT pinned by this suite — point readers to the legacy
  validate-plugin.test.ts for order-sensitive assertions).

validate:schemas + new error-codes lint green.
