+++ package.json
@@ -17,10 +17,11 @@
     "lint": "eslint . --ext .js,.ts",
     "test:unit": "vitest run --dir packages --passWithNoTests",
     "test:integration": "vitest run --dir tests/integration --passWithNoTests",
-    "validate:schemas": "node scripts/validate-marketplace.js && node scripts/validate-plugin.js && node scripts/validate-setup-all.js && node scripts/validate-agent-authoring.js",
+    "validate:schemas": "node scripts/validate-marketplace.js && node scripts/validate-plugin.js && node scripts/validate-setup-all.js && node scripts/validate-agent-authoring.js && node scripts/lint-error-codes.js",
     "validate:agents": "node scripts/validate-agent-authoring.js",
     "validate:marketplace": "node scripts/validate-marketplace.js",
     "validate:plugins": "node scripts/validate-plugin.js",
+    "validate:error-codes": "node scripts/lint-error-codes.js",
     "validate:setup-all": "node scripts/validate-setup-all.js",
     "validate:doc-counts": "node scripts/validate-doc-counts.js",
     "test:lint-plugins": "bats tests/lint-plugins.bats",
